feat: JWT authentication User & Middlewares

+++ api/controllers/users.controllers.js
@@ -1,6 +1,7 @@
 const User = require('../models/user.model');
 const mailer = require('../config/mailer.config');
 const createError = require('http-errors');
+const jwt = require('jsonwebtoken');
 
 module.exports.list = (req, res, next) => {
   User.find()
@@ -38,6 +39,10 @@ module.exports.detail = (req, res, next) => res.json(req.user);
 
 
 module.exports.update = (req, res, next) => {
+  if (req.user.id !== req.params.id) {
+    return next(createError(403, "Forbidden"))
+  }
+
   Object.assign(req.user, req.body);
   req.user.save()
     .then((user) => res.json(user))
@@ -47,6 +52,10 @@ module.exports.update = (req, res, next) => {
 
 
 module.exports.delete = (req, res, next) => {
+  if (req.user.id !== req.params.id) {
+    return next(createError(403, "Forbidden"))
+  }
+
   User.deleteOne({ _id: req.user.id })
     .then(() => res.status(204).send())
     .catch(next);
@@ -56,8 +65,12 @@ module.exports.delete = (req, res, next) => {
 module.exports.login = (req, res, next) => {
   User.findOne({ email: req.body.email })
     .then((user) => {
-      if (!user) {
+      if (!user || !req.body.password) {
         return next(createError(401, "Invalid credentials"));
+      }
+
+      if (!user.confirm) {
+        return next(createError(401, "Please confirm your account"));
       }
 
       user.checkPassword(req.body.password)
@@ -66,7 +79,8 @@ module.exports.login = (req, res, next) => {
             return next(createError(401, "Invalid credentials"));
           }
 
-          res.json(user);
+          const token = jwt.sign({ sub: user.id, exp: Date.now() / 1000 + 3_600 }, process.env.JWT_SECRET)
+          res.json({ token });
         });
     })
     .catch(next);
